Rename files upload_http.{h,cc} -> upload_gateway.{h,cc}
