+++ cvmfs/upload_gateway.h
