+++ test/unittests/t_gateway_uploader.cc
@@ -7,7 +7,7 @@
 #include <cstdlib>
 
 #include "upload.h"
-#include "upload_http.h"
+#include "upload_gateway.h"
 #include "upload_spooler_definition.h"
 #include "util/pointer.h"
 #include "util/posix.h"
